Time: 44 ms (82.99%), Space: 19 MB (45.12%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -9,7 +9,7 @@ public:
             
             int sum = 0;
             for (auto &i : piles) {
-                sum += (int)ceil((double)i / k);
+                sum += ceil((double)i / k);
                 if (sum > h) {
                     left = k + 1;
                     continue;
